fullstack/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
asyncpg
numpy
opencv-python-headless
python-multipart

#ML
torch
torchvision
tensorflow
Pillow
